Use PHP script to share functions
Also, constants script and functions script have own namespace
