+++ .tests/consts.php
@@ -1,8 +1,9 @@
 <?php
-namespace SuperUserDev\SchemaServer\Tests;
+namespace SuperUserDev\SchemaServer\Tests\Consts;
 const CREDS           = '../creds.json';
 const MIN_PHP_VERSION = '7.1';
 const DB_TEST         = true;
 const EXTS            = ['php'];
 const IGNORED_DIRS    = ['.git', '.tests'];
 const TZONE           = 'America/Los_Angeles';
+const CLI_SAPI        = ['cli', 'cli-server'];
